feat: add local file management functions

+++ client/src/Client.cpp
@@ -68,38 +68,44 @@ vector<string> Client::split_command(const string &command) {
 }
 
 void Client::process_command(const vector<string> &tokens) {
-
     if (tokens.empty()) return;
-    
+
     string command = tokens[0];
     transform(command.begin(), command.end(), command.begin(), ::tolower);
 
-    // TODO: checagem de erro TYPE ERROR
-    // funcao para cada bloco do if (handle_upload)
-    if (command == "upload" && tokens.size() == 2)
-        comm_manager.upload_file(tokens[1]);
-    else if (command == "download" && tokens.size() == 2)
-        comm_manager.download_file(tokens[1]);
-    else if (command == "delete" && tokens.size() == 2)
-        comm_manager.delete_file(tokens[1]);
-    else if (command == "list_server")
+    if (command == "upload" && tokens.size() == 2) {
+        if (file_manager.upload_local_file(tokens[1])) {
+            comm_manager.upload_file(tokens[1]);
+        }
+    } else if (command == "download" && tokens.size() == 2) {
+        if (file_manager.download_local_file(tokens[1])) {
+            comm_manager.download_file(tokens[1]);
+        }
+    } else if (command == "delete" && tokens.size() == 2) {
+        if (file_manager.delete_local_file(tokens[1])) {
+            comm_manager.delete_file(tokens[1]);
+    } else if (command == "list_server") {
         comm_manager.list_server();
-    else if (command == "list_client")
+
+    } else if (command == "list_client") {
         file_manager.list_files();
-    else if (command == "get_sync_dir")
+
+    } else if (command == "get_sync_dir") {
         std::cout << "Creating sync_dir and starting synchronization" << std::endl;
-    else if (command == "exit") {
+        // Pode chamar alguma função de criação aqui se necessário
+
+    } else if (command == "exit") {
         std::cout << "Closing session with server" << std::endl;
         comm_manager.exit_server();
-    }
-    else {
-        cout << "Invalid command or missing arguments. Available commands:" << endl;
-        cout << "# upload <path/filename.ext>" << endl;
-        cout << "# download <filename.ext>" << endl;
-        cout << "# delete <filename.ext>" << endl;
-        cout << "# list_server" << endl;
-        cout << "# list_client" << endl;
-        cout << "# get_sync_dir" << endl;
-        cout << "# exit" << endl;
+
+    } else {
+        std::cerr << "Invalid command or missing arguments. Available commands:" << std::endl;
+        std::cerr << "# upload <path/filename.ext>" << std::endl;
+        std::cerr << "# download <filename.ext>" << std::endl;
+        std::cerr << "# delete <filename.ext>" << std::endl;
+        std::cerr << "# list_server" << std::endl;
+        std::cerr << "# list_client" << std::endl;
+        std::cerr << "# get_sync_dir" << std::endl;
+        std::cerr << "# exit" << std::endl;
     }
 }
